Moved answers back to challenge directory
