+++ session1/3-challenge/1_arithmetic.rb
@@ -9,5 +9,5 @@
 # arithmetic1(-6)     # => -50
 
 def arithmetic1(n)
+  n * 5 - 20
 end
-
