+++ session1/3-challenge/2_arithmetic.rb
@@ -9,4 +9,9 @@
 # arithmetic2(-6, -7)  # => -3.5
 
 def arithmetic2(a, b)
+  if a < b
+    a/2.to_f
+  else
+    b/2.to_f
+  end
 end
