+++ session1/3-challenge/3_simple_logic.rb
@@ -8,4 +8,9 @@
 # ten_twenty(6) # => 10
 
 def ten_twenty(n)
+  if n % 2 == 0
+    10
+  else
+    20
+  end
 end
