+++ session1/3-challenge/4_logic.rb
@@ -13,6 +13,17 @@
 
 
 def grade(num_books, has_read_books)
+  if num_books < 10 && has_read_books == false
+    "D"
+  elsif num_books < 10 && has_read_books == true
+    "C"
+  elsif (num_books >= 10 && num_books <= 20) && has_read_books == false
+    "C"
+  elsif (num_books >= 10 && num_books <= 20) && has_read_books == true
+    "B"
+  elsif num_books > 20 && has_read_books == false
+    "B"
+  else 
+    "A"
+  end
 end
-
-
