--- session1/5-solved/1_arithmetic_solved.rb
@@ -1,13 +1,0 @@
-# fill out the method below
-# then test tho see if you did them correctly with
-#  $ rake 1:1
-
-# Given a number, return 20 less than, that number multiplied by 5
-#
-# arithmetic1(10)     # => 30
-# arithmetic1(10.5)   # => 32.5
-# arithmetic1(-6)     # => -50
-
-def arithmetic1(n)
-  n * 5 - 20
-end
