--- session1/5-solved/3_simple_logic_solved.rb
@@ -1,16 +1,0 @@
-# remember, you can test this file with
-#   $ rake 1:3
-
-
-# Given a number, n, return 10 if it is even, and 20 if it is odd
-#
-# ten_twenty(5) # => 20
-# ten_twenty(6) # => 10
-
-def ten_twenty(n)
-  if n % 2 == 0
-    10
-  else
-    20
-  end
-end
