--- session1/5-solved/5_string_solved.rb
@@ -1,9 +1,0 @@
-# Given a string, replace every instance of sad to happy
-#
-# add_more_ruby("The clowns were sad.")         # => "The clowns were happy."
-# add_more_ruby("The sad dad said sad stuff.")  # => "The happy dad said happy stuff."
-# add_more_ruby("Sad times are ahead!")         # => "Happy times are ahead!"
-
-def add_more_ruby(string)
-  string.gsub("sad", "happy").gsub("Sad", "Happy")
-end
