+++ session2/3-challenge/11_classes.rb
@@ -20,3 +20,53 @@
 # if the parameter is greater than 99, set the number of beer bottles to 99
 # Then make a public method called print_song that outputs all stanzas from the number of bottles of beer down to zero.
 # Add any additional methods you find helpful.
+
+# Create the clasee called beer song
+class BeerSong
+  # This creates the setter and getter methods (def beers=, def beers)
+  attr_accessor :beers
+
+  def initialize(beers)
+    # if beers is less that 0, beers = 0
+    beers = 0  if beers < 0
+    # if beer is > 99, beers = 99
+    beers = 99 if beers > 99
+    self.beers = beers
+  end
+
+  def print_song
+    # for beers, from 99 down to 1, print_stanza for each iteration
+    beers.downto 1 do |i|
+      print_stanza i
+    end
+  end
+
+  def print_stanza(n)
+    # if n = 0 make a new string
+    if n.zero?
+      String.new
+    else
+      # substitues the number into the song and takes one away at the end
+      puts "#{translate n} #{bottle n} of beer on the wall,"        ,
+           "#{translate n} #{bottle n} of beer,"                    ,
+           "Take one down, pass it around,"                         ,
+           "#{translate n - 1} #{bottle n-1} of beer on the wall."
+    end
+  end
+
+  # returns "bottle" or "bottles" depending if n = 1 or not
+  def bottle(n)
+    if n == 1 then 'bottle' else 'bottles' end
+  end
+
+  # translates number to English
+  def translate(n)
+    if 0 <= n && n <= 19
+      %w(zero one two three four five six seven eight nine ten eleven twelve thirteen fourteen fifteen sixteen seventeen eighteen nineteen)[n]
+    elsif n % 10 == 0
+      %w(zero ten twenty thirty forty fifty sixty seventy eighty ninety)[n/10]
+    else
+      "#{translate n/10*10}-#{translate n%10}".downcase
+    end.capitalize
+  end
+end
