+++ session2/3-challenge/2_input_output_control.rb
@@ -18,13 +18,46 @@
 # remember you can try your program out with              $ ruby 2_input_output_control.rb
 # and when you think it is correct, you can test it with  $ rake 2:2
 
-def hi_hi_goodbye
-  # your code here
-
+# A method that prompts the user for a number or bye
+def prompt
+  puts 'Enter a number or bye'
 end
 
+def hi_hi_goodbye
+  # Start off with the prompt
+  prompt
+  # While we gets a response and store it in line, and line doesn't include 'bye'
+  while (line = gets) && (line !~ /bye/)
+    # Convert line to an intiger, do what's in the block that many times
+    line.to_i.times { print 'hi ' }
+    # Then puts it (line)
+    puts
+    # Call prompt again
+    prompt
+  end
+  # Output Goodbye otherwise
+  puts "Goodbye!"
+end
 
+=begin Without comments:
 
+def prompt
+  puts 'Enter a number or bye'
+end
+
+def hi_hi_goodbye
+  prompt
+
+  while (line = gets) && (line !~ /bye/)
+    line.to_i.times { print 'hi ' }
+    puts
+    prompt
+  end
+
+  puts "Goodbye!"
+end
+
+=end
 
 # This will just invoke the method if you run this program directly
 # This way you can try it out by running "$ ruby 2_input_output_control.rb"
