+++ session2/3-challenge/8_array.rb
@@ -6,3 +6,25 @@
 # got_three? ['a', 'a', 'a']  # => true
 # got_three? [1, 2, 1, 1]     # => false
 
+def got_three?(elements)
+  # Iterate the following block for each array of consectuive 3 elements
+  elements.each_cons 3 do |a, b, c|
+    # if a = b and b = c, return true
+    return true if a == b && b == c
+  end
+  # otherwise return false
+  false
+end
+
+
+=begin Without comments:
+
+def got_three?(elements)
+  elements.each_cons 3 do |a, b, c|
+    return true if a == b && b == c
+  end
+  false
+end
+
+
+=end
