--- session2/5-solved/3_array_solved.rb
@@ -1,17 +1,0 @@
-# Write a method named every_other_char for strings that,
-# returns an array containing every other character
-#
-# example:
-# "abcdefg".every_other_char  # => "aceg"
-# "".every_other_char         # => ""
-
-class String
-  def every_other_char
-    to_return = ''
-    each_char.with_index do |char, index|
-      to_return << char if index.even?
-    end
-    to_return
-  end
-
-end
